requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.0
python-dotenv==1.0.1
pydantic==2.7.1
tavily-python==0.3.3
httpx==0.27.2
numpy>=1.24.0
